refactor: rename classes subpackage into core
