+++ quill/nn/_avgpool2d.py
@@ -2,7 +2,7 @@ from collections.abc import Sequence
 
 from . import Module
 from .functional import avgpool3d
-from ..classes import Tensor
+from ..core import Tensor
 from ..internals import type_check, len_check, expr_check
 
 class AvgPool2d(Module):
