+++ quill/nn/_linear.py
@@ -2,7 +2,7 @@ from cupy.random import rand
 
 from . import Module
 from .functional import addbias, expand_dims, matmul, squeeze
-from ..classes import Tensor
+from ..core import Tensor
 from ..internals import expr_check, type_check
 
 class Linear(Module):
