+++ quill/nn/_lstm.py
@@ -2,7 +2,7 @@ from cupy import zeros
 
 from . import Module, Linear
 from .functional import add, multiply, sigmoid, split, squeeze, stack, tanh
-from ..classes import Tensor
+from ..core import Tensor
 from ..internals import type_check, expr_check
 
 class _Cell(Module):
