+++ quill/nn/_module.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from ..classes import Tensor
+from ..core import Tensor
 
 _State = Tensor | dict[str, "_State"]
