+++ quill/nn/functional/_add.py
@@ -1,4 +1,4 @@
-from ...classes import Tensor
+from ...core import Tensor
 from ...internals import type_check
 
 def add(x1: Tensor, x2: Tensor) -> Tensor:
