+++ quill/nn/functional/_concatenate.py
@@ -2,7 +2,7 @@ from collections.abc import Sequence
 from cupy import concatenate as _concatenate, split
 from numpy import ndarray
 
-from ...classes import Tensor
+from ...core import Tensor
 from ...internals import type_check
 
 def concatenate(tensors: Sequence[Tensor], axis: int = 0) -> Tensor:
