+++ quill/nn/functional/_conv3d.py
@@ -2,7 +2,7 @@ from collections.abc import Sequence
 from cupy import flip, pad, sum, zeros
 from numpy import ndarray
 
-from ...classes import Tensor
+from ...core import Tensor
 from ...internals import expr_check, len_check, type_check
 
 def conv3d(input_tensor: Tensor, kernel: Tensor, stride: int | Sequence[int] = 1, padding: int | Sequence[int] = 0) -> Tensor:
