+++ quill/nn/functional/_matmul.py
@@ -1,7 +1,7 @@
 from numpy import ndarray
 from cupy import matmul as _matmul, sum, swapaxes
 
-from ...classes import Tensor
+from ...core import Tensor
 from ...internals import type_check
 
 def matmul(x1: Tensor, x2: Tensor) -> Tensor:
