+++ quill/nn/functional/_split.py
@@ -1,6 +1,6 @@
 from cupy import split as _split
 
-from ...classes import Tensor
+from ...core import Tensor
 from ...internals import type_check
 
 def split(tensor: Tensor, indices_or_sections: int, axis: int = 0) -> list[Tensor]:
