+++ quill/nn/functional/_squeeze.py
@@ -1,6 +1,6 @@
 from cupy import expand_dims, squeeze as _squeeze
 
-from ...classes import Tensor
+from ...core import Tensor
 from ...internals import type_check
 
 def squeeze(x: Tensor, axis: int) -> Tensor:
