+++ quill/nn/functional/_stack.py
@@ -2,7 +2,7 @@ from collections.abc import Sequence
 from cupy import split, squeeze, stack as _stack
 from numpy import ndarray
 
-from ...classes import Tensor
+from ...core import Tensor
 from ...internals import type_check
 
 def stack(tensors: Sequence[Tensor], axis: int = 0) -> Tensor:
