+++ quill/nn/functional/_tanh.py
@@ -1,7 +1,7 @@
 from cupy import exp
 from numpy import ndarray
 
-from ...classes import Tensor
+from ...core import Tensor
 from ...internals import type_check
 
 def tanh(x: Tensor) -> Tensor:
